Merge https://github.com/futuretap/InAppSettingsKit's master
and some patches cocologics still relies on.

+++ InAppSettingsKit/Models/IASKSettingsStore.h
@@ -24,12 +24,12 @@
 - (void)setBool:(BOOL)value      forKey:(NSString*)key;
 - (void)setFloat:(float)value    forKey:(NSString*)key;
 - (void)setDouble:(double)value  forKey:(NSString*)key;
-- (void)setInteger:(int)value    forKey:(NSString*)key;
+- (void)setInteger:(NSInteger)value    forKey:(NSString*)key;
 - (void)setObject:(id)value      forKey:(NSString*)key;
 - (BOOL)boolForKey:(NSString*)key;
 - (float)floatForKey:(NSString*)key;
 - (double)doubleForKey:(NSString*)key;
-- (int)integerForKey:(NSString*)key;
+- (NSInteger)integerForKey:(NSString*)key;
 - (id)objectForKey:(NSString*)key;
 - (BOOL)synchronize; // Write settings to a permanant storage. Returns YES on success, NO otherwise
 @end
